[SMTC-1694] Add extension functions for generating minutes and hours from ints

+++ app/src/main/java/com/ciandt/dragonfly/example/DragonflyApplication.kt
@@ -96,7 +96,7 @@ class DragonflyApplication : Application() {
         firebaseRemoteConfig.setDefaults(R.xml.remote_config_defaults)
 
         firebaseRemoteConfig
-                .fetch(FirebaseConfig.REMOTE_CONFIG_CACHE_EXPIRATON)
+                .fetch(FirebaseConfig.REMOTE_CONFIG_CACHE_EXPIRATON.toLong())
                 .addOnCompleteListener { task ->
                     if (task.isSuccessful) {
                         firebaseRemoteConfig.activateFetched()
